Make the approximation shader more efficient.
Instead of computing all of the root products in the shader,
precompute the polynomial resulting from these products on the CPU
and pass the coefficients to the shader. The approximation routine
is then simply computing the value of two polynomials at a given point
and dividing, in exactly the same fashion as the original f/f'
polynomial division approach.